Improved type coverage

+++ src/php/Lang/PhelArray.php
@@ -10,6 +10,10 @@ use InvalidArgumentException;
 use Iterator;
 use Phel\Printer\Printer;
 
+/**
+ * @template-implements ArrayAccess<int, mixed>
+ * @template-implements Iterator<int, mixed>
+ */
 final class PhelArray extends AbstractType implements
     ArrayAccess,
     Countable,
@@ -22,6 +26,7 @@ final class PhelArray extends AbstractType implements
     PushInterface,
     ConcatInterface
 {
+    /** @var mixed[] */
     private array $data;
 
     /**
@@ -84,7 +89,7 @@ final class PhelArray extends AbstractType implements
     }
 
     /**
-     * @param mixed $offset
+     * @param int $offset
      *
      * @return mixed|null
      */
@@ -98,6 +103,9 @@ final class PhelArray extends AbstractType implements
         return count($this->data);
     }
 
+    /**
+     * @return mixed
+     */
     public function current()
     {
         return current($this->data);
@@ -105,7 +113,7 @@ final class PhelArray extends AbstractType implements
 
     public function key()
     {
-        return key($this->data);
+        return (int) key($this->data);
     }
 
     public function next(): void
@@ -192,6 +200,7 @@ final class PhelArray extends AbstractType implements
 
     public function concat($xs): ConcatInterface
     {
+        /** @var mixed $x */
         foreach ($xs as $x) {
             $this->data[] = $x;
         }
